services/patch_differ/requirements.txt
karton-core
mwdblib
pefile
requests
PyYAML>=6.0
